pyproject.toml
[tool.poetry]
name = "user microservice"
version = "0.1"
description = "User microservice"
authors = ["LedxDeliveryFlopp <korstim18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.0"}
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
httpx = "^0.27.0"
passlib = "^1.7.4"
ruff = "^0.3.7"
pre-commit = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"


[tool.ruff]
exclude = ["migrations", "usr", "venv", "tmp", "var"]